libsubprocess: move all server code to server.c
Problem: subprocess server implementation is split between
subprocess.c and server.c.

Consolidate in server.c.  There are no other code changes in this
commit, just relocation.
