+++ src/common/libsubprocess/subprocess.h
@@ -35,9 +35,6 @@ typedef struct flux_command flux_cmd_t;
  */
 typedef struct flux_subprocess flux_subprocess_t;
 
-/*  flux_subprocess_server_t: Handler for a subprocess remote server */
-typedef struct flux_subprocess_server flux_subprocess_server_t;
-
 /*
  * Subprocess states, on changes, will lead to calls to
  * on_state_change below.
@@ -112,53 +109,6 @@ typedef struct {
     flux_subprocess_hook_f post_fork;
     void *post_fork_arg;
 } flux_subprocess_hooks_t;
-
-/*
- *  General support:
- */
-
-/*  Start a subprocess server on the handle `h`. Registers message
- *   handlers, etc for remote execution.
- */
-flux_subprocess_server_t *flux_subprocess_server_start (flux_t *h,
-                                                        const char *local_uri,
-                                                        uint32_t rank);
-
-
-typedef int (*flux_subprocess_server_auth_f) (const flux_msg_t *msg,
-                                              void *arg);
-
-/*   Register an authorization function to the subprocess server
- *
- *   The registered function should return 0 to allow the request to
- *    proceed, and -1 with errno set to deny the request.
- */
-void flux_subprocess_server_set_auth_cb (flux_subprocess_server_t *s,
-                                         flux_subprocess_server_auth_f fn,
-                                         void *arg);
-
-/*  Stop a subprocess server / cleanup flux_subprocess_server_t.  Will
- *  send a SIGKILL to all remaining subprocesses.
- */
-void flux_subprocess_server_stop (flux_subprocess_server_t *s);
-
-/* Send all subprocesses signal and wait up to wait_time seconds for
- * all subprocesses to complete.  This is typically called to send
- * SIGTERM before calling flux_subprocess_server_stop(), allowing
- * users to send a signal to inform subprocesses to complete / cleanup
- * before they are sent SIGKILL.
- *
- * This function will enter the reactor to wait for subprocesses to
- * complete, should only be called on cleanup path when primary
- * reactor has exited.
- */
-int flux_subprocess_server_subprocesses_kill (flux_subprocess_server_t *s,
-                                              int signum,
-                                              double wait_time);
-
-/* Terminate all subprocesses started by a sender id */
-int flux_subprocess_server_terminate_by_uuid (flux_subprocess_server_t *s,
-                                              const char *id);
 
 /*
  * Convenience Functions:
